Update Incremental Model - Alvesm.txt
There was a comma issue in the third paragraph that I fixed. The issue with this was that Matt had a double comma. I had to remove it. I read over the entire assignment, and I thought it was very nice, and very detailed about what each of the students would have to do regarding the CRN's